LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-pannernode-interface/panner-equalpower-expected.txt

CONSOLE MESSAGE: line 70: TypeError: undefined is not an object (evaluating 'context.startRendering().then')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'context.startRendering().then')

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [test] Equal-power panner model of AudioPannerNode 
